add filter with jobNames to myPositionsPage (without filter functionality)
