+++ src/main/webapp/i18n/de/job.json
@@ -128,7 +128,10 @@
       "createButton": "Neue Stelle erstellen"
     },
     "searchFilterSortBar": {
-      "searchText": "Stellen suchen..."
+      "searchText": "Stellen suchen...",
+      "filterOptions": {
+        "jobSearchPlaceholder": "Suche Stellen..."
+      }
     },
     "tableColumn": {
       "supervisingProfessor": "Betreuender Professor",
@@ -150,6 +153,9 @@
       "closeMessage": "Bist du sicher, dass du diese Position schließen möchtest? Nach dem Schließen können sich Bewerber nicht mehr auf diese Position bewerben. Diese Aktion kann NICHT rückgängig gemacht werden.",
       "deleteHeader": "Löschvorgang bestätigen",
       "deleteMessage": "Bist du sicher, dass du diesen Entwurf dauerhaft löschen möchtest? Diese Aktion kann NICHT rückgängig gemacht werden!"
+    },
+    "errors": {
+      "loadJobNames": "Fehler beim Laden der Jobnamen. Bitte versuche, die Seite neu zu laden, oder kontaktiere uns, wenn das Problem weiterhin besteht."
     }
   },
   "jobDetailPage": {
